refactor: extract bp_sessions to routes/sessions.py + update CLAUDE.md
First step of incremental modularisation. Pure mechanical move — zero
behaviour change, zero new endpoints. Validated via local dashboard
boot + curl spot-check + browser E2E (MODEL/TOKENS/SESSIONS/spend
all match ground truth post-refactor).

## What moved

All 13 routes registered on `bp_sessions`:

  /api/sessions
  /api/compactions
  /api/session-tools
  /api/cost-split
  /api/task-runs
  /api/subagents
  /api/delegation-tree
  /api/export/otlp
  /api/sessions/cost-breakdown
  /api/sessions/<session_id>/stop
  /api/transcripts
  /api/transcript/<session_id>
  /api/transcript-events/<session_id>

`bp_sessions` blueprint definition + all 13 handlers now live in
`routes/sessions.py` (1,187 lines). `dashboard.py` shrank by ~1,100
lines (34,297 → 33,193) and gains a single import:

    from routes.sessions import bp_sessions

at the top, alongside the other blueprint registrations.

## How shared helpers stay accessible

Each handler does a late `import dashboard as _d` at the top of its
body and references shared helpers as `_d._get_sessions()`,
`_d._compute_transcript_analytics()`, etc. This avoids circular
imports while keeping `dashboard.py` as the single source of truth
for module-level state, helpers, and the embedded HTML/CSS/JS.

Pattern matches the existing `clawmetry-cloud/routes/cloud.py`
convention.

## Why this matters (fixes the recurring conflict pattern)

Recent PRs (#613 #614 #615 #616 #617 #618 #619 #620) all wanted to
add new endpoints in the same spot in dashboard.py — right after
`api_sessions`. They kept conflicting with each other on rebase.
With per-feature route modules, parallel PRs adding new endpoints
touch their own feature's file and naturally don't conflict.

## CLAUDE.md update

Removed the "single file — don't split dashboard.py" rule and
documented the new convention. Original rationale (portability,
auditability) made sense at small scale but inverted at ~33K lines.
File listing updated to reflect new `routes/` directory.

## Validated (live, pre-push)

  python3 -c "import dashboard"               → ok
  python3 dashboard.py --port 8910            → boots cleanly
  curl /api/sessions, /api/compactions,
       /api/session-tools, /api/cost-split,
       /api/task-runs, /api/subagents,
       /api/delegation-tree,
       /api/sessions/cost-breakdown,
       /api/transcripts                       → all 200
  Browser dashboard at localhost:8910         → renders, MODEL
       badge, TOKENS, SESSIONS count, $$ all match ground truth

## Future work (separate PRs)

Same pattern applied to `bp_brain`, `bp_usage`, `bp_crons`,
`bp_health`, etc. Each is an independent move with no behaviour
change.
